simplify the book screen, and delete the shopping cart related
