--- shared/src/commonMain/kotlin/com/dishdiscoverers/foodrecipe/dongguo/layout/MyTopBar.kt
@@ -1,20 +1,0 @@
-package com.dishdiscoverers.foodrecipe.dongguo.layout
-
-import androidx.compose.material.Text
-import androidx.compose.material3.CenterAlignedTopAppBar
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.TopAppBarScrollBehavior
-import androidx.compose.runtime.Composable
-
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun MyTopBar(
-    message: String,
-    scrollBehavior: TopAppBarScrollBehavior
-) {
-    CenterAlignedTopAppBar(
-        scrollBehavior = scrollBehavior,
-        title = { Text(text = message) },
-    )
-}
